@@ -9,7 +9,7 @@ repositories {
     mavenCentral()
 }
 
-sourceCompatibility = '1.8'
+sourceCompatibility = '1.17'
 
 tasks.withType(JavaCompile) {
     options.encoding = 'UTF-8'
