@@ -1,17 +1,27 @@
 package com.symphony.bdk.examples.app.spring;
 
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.web.SecurityFilterChain;
 
-@EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+@Configuration
+public class SecurityConfig {
 
-  @Override
-  protected void configure(HttpSecurity http) throws Exception {
+  @Bean
+  public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
     // This configuration is not recommended in production setup
     http.authorizeRequests().anyRequest().permitAll();
     http.csrf().disable();
     http.headers().frameOptions().disable();
+    return http.build();
   }
+
+  @Bean
+  public WebSecurityCustomizer webSecurityCustomizer() {
+    //TODO: Fix this bean method
+    return null;
+  }
+
 }
