@@ -2,16 +2,12 @@ package com.symphony.bdk.app.spring.auth.model;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
 
 /**
  * App Token generated by the application.
  */
 
-@Data
-public class AppToken {
-
-  private final String appToken;
+public record AppToken(String appToken) {
 
   @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
   public AppToken(@JsonProperty("appToken") String appToken) {
